Adds docs for @codegen.function decorator (#187)
#Motivation
<!-- Why is this change necessary? -->
To provide clear documentation on how to use the function decorator
within the codegen SDK, including examples of deploying and running
functions using the CLI.
#Content
<!-- Please include a summary of the change -->
Added a new document function-decorator.mdx in the building-with-codegen
directory. This document includes usage instructions, examples, and CLI
commands for deploying and running functions.
#Testing

<!-- How was the change tested? -->
Reviewed the documentation for accuracy and clarity. Verified that the
links to related documents are correct.
Please check the following before marking your PR as ready for review
[x] I have added tests for my changes
[x] I have updated the documentation or added new documentation as
needed
[x] I have read and agree to the Contributor License Agreement
---